+++ tests/date_tests/test_diff.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from datetime import date
-from pendulum import Date, Pendulum
+from pendulum import Pendulum, Date
 
 from .. import AbstractTestCase
 
@@ -123,8 +123,12 @@ class DiffTest(AbstractTestCase):
         self.assertEqual('3 weeks ago', Date.today().subtract(weeks=3).diff_for_humans())
 
     def test_diff_for_humans_now_and_month(self):
-        self.assertEqual('4 weeks ago', Date.today().subtract(weeks=4).diff_for_humans())
-        self.assertEqual('1 month ago', Date.today().subtract(months=1).diff_for_humans())
+        with self.wrap_with_test_now(Pendulum.create(2016, 3, 1)):
+            self.assertEqual('4 weeks ago', Date.today().subtract(weeks=4).diff_for_humans())
+            self.assertEqual('1 month ago', Date.today().subtract(months=1).diff_for_humans())
+
+        with self.wrap_with_test_now(Pendulum.create(2017, 2, 28)):
+            self.assertEqual('1 month ago', Date.today().subtract(weeks=4).diff_for_humans())
 
     def test_diff_for_humans_now_and_months(self):
         self.assertEqual('2 months ago', Date.today().subtract(months=2).diff_for_humans())
@@ -159,6 +163,12 @@ class DiffTest(AbstractTestCase):
     def test_diff_for_humans_now_and_future_month(self):
         with self.wrap_with_test_now(Pendulum.create(2016, 3, 1)):
             self.assertEqual('4 weeks from now', Date.today().add(weeks=4).diff_for_humans())
+            self.assertEqual('1 month from now', Date.today().add(months=1).diff_for_humans())
+
+        with self.wrap_with_test_now(Pendulum.create(2017, 3, 31)):
+            self.assertEqual('1 month from now', Date.today().add(months=1).diff_for_humans())
+
+        with self.wrap_with_test_now(Pendulum.create(2017, 4, 30)):
             self.assertEqual('1 month from now', Date.today().add(months=1).diff_for_humans())
 
         with self.wrap_with_test_now(Pendulum.create(2017, 1, 31)):
@@ -199,6 +209,12 @@ class DiffTest(AbstractTestCase):
             self.assertEqual('4 weeks before', Date.today().diff_for_humans(Date.today().add(weeks=4)))
             self.assertEqual('1 month before', Date.today().diff_for_humans(Date.today().add(months=1)))
 
+        with self.wrap_with_test_now(Pendulum.create(2017, 3, 31)):
+            self.assertEqual('1 month before', Date.today().diff_for_humans(Date.today().add(months=1)))
+
+        with self.wrap_with_test_now(Pendulum.create(2017, 4, 30)):
+            self.assertEqual('1 month before', Date.today().diff_for_humans(Date.today().add(months=1)))
+
         with self.wrap_with_test_now(Pendulum.create(2017, 1, 31)):
             self.assertEqual('1 month before', Date.today().diff_for_humans(Date.today().add(weeks=4)))
 
@@ -233,8 +249,12 @@ class DiffTest(AbstractTestCase):
         self.assertEqual('3 weeks after', Date.today().diff_for_humans(Date.today().subtract(weeks=3)))
 
     def test_diff_for_humans_other_and_future_month(self):
-        self.assertEqual('4 weeks after', Date.today().diff_for_humans(Date.today().subtract(weeks=4)))
-        self.assertEqual('1 month after', Date.today().diff_for_humans(Date.today().subtract(months=1)))
+        with self.wrap_with_test_now(Pendulum.create(2016, 3, 1)):
+            self.assertEqual('4 weeks after', Date.today().diff_for_humans(Date.today().subtract(weeks=4)))
+            self.assertEqual('1 month after', Date.today().diff_for_humans(Date.today().subtract(months=1)))
+
+        with self.wrap_with_test_now(Pendulum.create(2017, 2, 28)):
+            self.assertEqual('1 month after', Date.today().diff_for_humans(Date.today().subtract(weeks=4)))
 
     def test_diff_for_humans_other_and_future_months(self):
         self.assertEqual('2 months after', Date.today().diff_for_humans(Date.today().subtract(months=2)))
@@ -263,6 +283,19 @@ class DiffTest(AbstractTestCase):
     def test_diff_for_humans_absolute_years(self):
         self.assertEqual('1 year', Date.today().diff_for_humans(Date.today().subtract(years=1), True))
         self.assertEqual('1 year', Date.today().diff_for_humans(Date.today().add(years=1), True))
+
+    def test_diff_for_humans_accuracy(self):
+        today = Pendulum.today()
+
+        with self.wrap_with_test_now(today.add(days=1)):
+            self.assertEqual('1 year', today.add(years=1).diff_for_humans(absolute=True))
+
+        with self.wrap_with_test_now(today):
+            self.assertEqual('6 days', today.add(days=6).diff_for_humans(absolute=True))
+            self.assertEqual('1 week', today.add(days=7).diff_for_humans(absolute=True))
+            self.assertEqual('3 weeks', today.add(days=20).diff_for_humans(absolute=True))
+            self.assertEqual('2 weeks', today.add(days=14).diff_for_humans(absolute=True))
+            self.assertEqual('2 weeks', today.add(days=13).diff_for_humans(absolute=True))
 
     def test_subtraction(self):
         d = Date(2016, 7, 5)