+++ tests/localization_tests/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import pendulum
 from pendulum import Pendulum
 
 
@@ -8,13 +9,13 @@ class AbstractLocalizationTestCase(object):
     locale = None
 
     def setUp(self):
-        Pendulum.set_locale(self.locale)
+        pendulum.set_locale(self.locale)
 
     def tearDown(self):
-        Pendulum.set_locale('en')
+        pendulum.set_locale('en')
 
     def test_diff_for_humans_localized(self):
-        with Pendulum.test(Pendulum(2016, 8, 29)):
+        with pendulum.test(Pendulum(2016, 8, 29)):
             self.diff_for_humans()
 
     def diff_for_humans(self):